requirements.txt
aniso8601==9.0.1
annotated-types==0.7.0
anyio==4.4.0
asttokens==2.4.1
blinker==1.8.2
cached-property==1.5.2
cachetools==5.4.0
certifi==2024.7.4
cffi==1.17.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
cowsay==6.1
cryptography==43.0.0
cx_Oracle==8.3.0
decorator==5.1.1
distlib==0.3.8
distro==1.9.0
et-xmlfile==1.1.0
executing==2.0.1
filelock==3.15.4
Flask==3.0.3
Flask-RESTful==0.3.10
fsspec==2024.6.1
gcloud==0.18.3
google-ai-generativelanguage==0.6.6
google-api-core==2.19.1
google-api-python-client==2.139.0
google-auth==2.32.0
google-auth-httplib2==0.2.0
google-generativeai==0.7.2
googleapis-common-protos==1.63.2
greenlet==3.0.3
grpcio==1.65.4
grpcio-status==1.62.2
h11==0.14.0
httpcore==1.0.5
httplib2==0.22.0
httpx==0.27.0
huggingface-hub==0.24.5
idna==3.7
inflection==0.5.1
ipython==8.26.0
itsdangerous==2.2.0
jedi==0.19.1
Jinja2==3.1.4
jwcrypto==1.5.6
Markdown==3.6
MarkupSafe==2.1.5
matplotlib-inline==0.1.7
mpmath==1.3.0
mypy-extensions==1.0.0
networkx==3.3
numpy==2.0.1
oauth2client==4.1.3
openai==1.37.1
openpyxl==3.1.5
packaging==24.1
pandas==2.2.2
parso==0.8.4
pillow==10.4.0
pipenv==2024.0.1
platformdirs==4.2.2
prompt_toolkit==3.0.47
proto-plus==1.24.0
protobuf==4.25.4
pure_eval==0.2.3
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
pycryptodome==3.20.0
pydantic==2.8.2
pydantic_core==2.20.1
Pygments==2.18.0
pyparsing==3.1.2
pypng==0.20220715.0
Pyrebase4==4.8.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-jwt==4.1.0
pytz==2024.1
PyYAML==6.0.2
qrcode==7.4.2
regex==2024.7.24
requests==2.32.3
requests-toolbelt==0.10.1
rsa==4.9
safetensors==0.4.4
setuptools==72.1.0
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.31
sqlalchemy-orm==1.2.10
stack-data==0.6.3
sympy==1.13.2
tokenizers==0.19.1
torch==2.4.0
tqdm==4.66.4
traitlets==5.14.3
transformers==4.44.0
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2024.1
uritemplate==4.1.1
urllib3==1.26.19
virtualenv==20.26.3
wcwidth==0.2.13
Werkzeug==3.0.3
